src/requirements.txt
psycopg2_binary
lxml
requests
fastapi
python-dotenv
starlette
uvicorn
